Handle errors in the build script

+++ scripts/buildAndroid.sh
@@ -26,7 +26,6 @@ TIMESTAMP=`date +%Y%m%d`
 # setup some enviroment variables
 export ANDROID_NDK_HOST=linux-x86_64
 export ANDROID_HOME=$ANDROID_SDK_ROOT
-export ANDROID_NDK_PLATFORM=$NDK_PLATFORM_32BIT
 export JAVA_HOME=$JDK
 
 #
@@ -42,30 +41,48 @@ QMAKE=$QT_DIR/android_armv7/bin/qmake
 # call qmake
 echo ">> $QMAKE -o Makefile $SOURCE_DIR/src/src.pro -spec android-clang CONFIG+=qtquickcompiler"
 $QMAKE $SOURCE_DIR/src/src.pro -o Makefile -spec android-clang CONFIG+=qtquickcompiler
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 
 MAKE=$ANDROID_NDK_ROOT/prebuilt/linux-x86_64/bin/make
 # build & install
 echo ">> $MAKE -f Makefile"
 $MAKE -f Makefile
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 echo ">> $MAKE INSTALL_ROOT=$BUILD_DIR/android-build -f Makefile install"
 $MAKE INSTALL_ROOT=$BUILD_DIR/android-build -f Makefile install
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 # create the apk
 DEPLOY_TOOL=$QT_DIR/android_armv7/bin/androiddeployqt
 SETTING=$BUILD_DIR/android-libBaugeschichte.so-deployment-settings.json
 echo ">> $DEPLOY_TOOL --input $SETTING --output $BUILD_DIR/android-build --android-platform $PLATFORM --jdk $JDK --gradle --release"
 $DEPLOY_TOOL --input $SETTING --output $BUILD_DIR/android-build --android-platform $PLATFORM --jdk $JDK --gradle --release
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 
 # copy file
 APK32=$BUILD_DIR/android-build/build/outputs/apk/android-build-release-unsigned.apk
 DESTINATION32=$TARGET_DIR/Baugeschichte_signed_arm32_$TIMESTAMP.apk
 cp $APK32 $DESTINATION32
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 
 
 #
 # build ARM 64 bit
 #
-export ANDROID_NDK_PLATFORM=$NDK_PLATFORM_64BIT
- 
 BUILD_DIR=$BUILD_BASE/arm64
 rm -rf $BUILD_DIR
 mkdir -p $BUILD_DIR
@@ -75,23 +92,44 @@ QMAKE=$QT_DIR/android_arm64_v8a/bin/qmake
 # call qmake
 echo ">> $QMAKE -o Makefile $SOURCE_DIR/src/src.pro -spec android-clang CONFIG+=qtquickcompiler"
 $QMAKE -o Makefile $SOURCE_DIR/src/src.pro -spec android-clang CONFIG+=qtquickcompiler
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
  
 MAKE=$ANDROID_NDK_ROOT/prebuilt/linux-x86_64/bin/make
 # build & install
 echo ">> $MAKE -f Makefile"
 $MAKE -f Makefile
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 echo ">> $MAKE INSTALL_ROOT=$BUILD_DIR/android-build -f Makefile install"
 $MAKE INSTALL_ROOT=$BUILD_DIR/android-build -f Makefile install
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
+
 # create the apk
 DEPLOY_TOOL=$QT_DIR/android_arm64_v8a/bin/androiddeployqt
 SETTING=$BUILD_DIR/android-libBaugeschichte.so-deployment-settings.json
 echo ">> $DEPLOY_TOOL --input $SETTING --output $BUILD_DIR/android-build --android-platform $PLATFORM --jdk $JDK --gradle --release"
 $DEPLOY_TOOL --input $SETTING --output $BUILD_DIR/android-build --android-platform $PLATFORM --jdk $JDK --gradle --release
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 
 # copy file
 APK64=$BUILD_DIR/android-build/build/outputs/apk/android-build-release-unsigned.apk
 DESTINATION64=$TARGET_DIR/Baugeschichte_signed_arm64_$TIMESTAMP.apk
 cp $APK64 $DESTINATION64
+if [ $? -ne 0 ]; then
+  echo "Error building Baugeschichte"
+  exit 1
+fi
 
 
 echo " "
@@ -99,3 +137,4 @@ echo "Copied result to $DESTINATION32"
 echo "Copied result to $DESTINATION64"
 echo " "
  
+exit 0
